requirements.txt
deta==1.1.0
gensim==4.3.1
IMDbPY==2022.7.9
IMDbPY==2022.7.9
nltk==3.8.1
numpy==1.23.5
pandas==1.5.3
python-dotenv==1.0.0
pytube==12.1.3
Requests==2.30.0
scikit_learn==1.2.2
streamlit==1.21.0
streamlit_player==0.1.5
tmdbv3api==1.7.7
